+++ resources/query.properties
@@ -10,3 +10,4 @@ selectAllLikes=SELECT * FROM LIKES WHERE user_id = ?;
 selectByDistrict=SELECT * FROM travels WHERE district = ?;
 editPassword=UPDATE USERS SET password = ?, updated_at = CURRENT_TIMESTAMP WHERE user_id = ?;
 editName=UPDATE USERS SET name = ?, updated_at = CURRENT_TIMESTAMP WHERE user_id = ?;
+findById=SELECT user_id, role, name, password, created_at, updated_at FROM users WHERE user_id = ?;
