+++ src/com/multi/controller/LikeController.java
@@ -5,6 +5,7 @@ import com.multi.model.dto.LikesDTO;
 import com.multi.model.dto.TravelDTO;
 import com.multi.model.dto.tmddk.Like;
 import com.multi.service.LikeService;
+import com.multi.view.DisplayView;
 import com.multi.view.GeneralView;
 import com.multi.view.MainMenu;
 
@@ -13,6 +14,8 @@ import java.util.ArrayList;
 
 
 public class LikeController {
+    private static DisplayView displayView = new DisplayView();
+
     private LikeService likeService = new LikeService();
     private GeneralView generalView = new GeneralView();
     public void selectAllLikes(String userId) {
@@ -41,11 +44,11 @@ public class LikeController {
     public void insertLike(LikesDTO like) {
         int result = likeService.insertLike(like);
         if (result > 0) {
-            generalView.displayMessage(">> 즐겨찾기에 성공적으로 등록되었습니다.");
+            displayView.displayMessage(">> 즐겨찾기에 성공적으로 등록되었습니다.");
         } else if (result == -1) {
-            generalView.displayError(">> 이미 등록된 즐겨찾기입니다."); // displayError 메소드가 없다면 System.out.println() 사용
+            displayView.displayError(">> 이미 등록된 즐겨찾기입니다."); // displayError 메소드가 없다면 System.out.println() 사용
         } else {
-            generalView.displayError(">> 즐겨찾기 등록에 실패했습니다. 다시 시도해주세요.");
+            displayView.displayError(">> 즐겨찾기 등록에 실패했습니다. 다시 시도해주세요.");
         }
     }
