+++ src/com/multi/service/EditProfileService.java
@@ -1,6 +1,7 @@
 package com.multi.service;
 
 import com.multi.model.dao.UserDAO;
+import com.multi.model.dto.tmddk.User;
 
 import java.sql.Connection;
 
@@ -38,4 +39,24 @@ public class EditProfileService {
         close(conn);
         return result;
     }
+
+    public User editNameAndRefresh(String newName) {
+        Connection conn = null;
+        try {
+            conn = getConnection();
+            int updated = userDAO.editName(conn, newName);
+            if (updated > 0) {
+                // DB에서 최신값 재조회
+                String uid = UserSession.getUser().getUserId();
+                User fresh = userDAO.findById(conn, uid);
+                // 세션 갱신
+                UserSession.setUser(fresh);
+                return fresh;
+            }
+            return null;
+        } finally {
+            close(conn);
+        }
+    }
+
 }
