+++ src/com/multi/view/GeneralView.java
@@ -1,11 +1,8 @@
 package com.multi.view;
 
 import com.multi.controller.GeneralController;
-import com.multi.controller.TravelDetailController;
-import com.multi.model.dto.TravelDTO;
 import com.multi.model.dto.tmddk.User;
 
-import java.util.ArrayList;
 import java.util.InputMismatchException;
 import java.util.Scanner;
 
@@ -14,8 +11,8 @@ public class GeneralView {
     Scanner scanner = new Scanner(System.in);
     GeneralController generalController = new GeneralController();
     private static GeneralView generalView = new GeneralView();
+    private static DisplayView displayView = new DisplayView();
     private static MainMenu mainMenu = new MainMenu();
-    private TravelDetailController travelDetailController = new TravelDetailController();
     public void signUpView() {
         generalController.signUp(inputUser());
     }
@@ -62,22 +59,14 @@ public class GeneralView {
 
         User user = generalController.checkGeneral(userId, password);
         if (user != null) {
-            displayMessage("로그인 성공!");
+            displayView.displayMessage("로그인 성공!");
             mainMenu.mainMenu();
         } else {
-            displayMessage("아이디 또는 비밀번호가 올바르지 않습니다.");
+            displayView.displayMessage("아이디 또는 비밀번호가 올바르지 않습니다.");
         }
     }
 
 
-    public void displayMessage(String s) {
-
-        System.out.println("서비스 요청결과  : " + s);
-
-    }
-    public void displayError(String message) {
-        System.out.println( message);
-    }
 
     public void loginOrSignup() {
         int choice = 0;
@@ -112,148 +101,5 @@ public class GeneralView {
 
     }
 
-
-    public void displayTravel(ArrayList<TravelDTO> dto) {
-
-        final int PAGE_SIZE = 20;
-        // Math.ceil을 사용하기 위해 double로 형변환하여 정확한 페이지 수 계산
-        int totalPages = (int) Math.ceil((double) dto.size() / PAGE_SIZE);
-        int currentPage = 0;
-
-
-        if (dto == null || dto.isEmpty()) {
-            System.out.println("조회된 여행지 정보가 없습니다.");
-            return;
-        }
-        // 페이지네이션 제어 루프
-        while (true) {
-            System.out.println("\n\n");
-            System.out.println("조회된 여행지 리스트는 다음과 같습니다.");
-
-            //데이터 슬라이싱 및 표시 로직
-
-            int startIndex = currentPage * PAGE_SIZE;
-            int endIndex = Math.min(startIndex + PAGE_SIZE, dto.size());
-            for (int i = startIndex; i < endIndex; i++) {
-
-                System.out.println(dto.get(i));
-
-            }
-
-
-            //UI 요소
-
-            System.out.println("\n--- Page " + (currentPage + 1) + " of " + totalPages + " ---");
-            System.out.print("\n[p]revious, [n]ext, [h]ome | 여행지 상세정보는 번호 입력: ");
-
-            String command = sc.next();
-
-            try {
-                int travelNo = Integer.parseInt(command);
-
-                // 2. 숫자로 변환 성공 시, 상세 정보 조회 로직 실행
-                TravelDTO selectedTravel = travelDetailController.showDetail(travelNo);
-
-                if (selectedTravel != null) {
-                    // 상세 정보를 출력하는 메소드 호출
-                    displayTravelDetail(selectedTravel);
-
-
-                    boolean inDetailView = true;
-                    while (inDetailView) {
-                        System.out.println("\n");
-                        System.out.println("목록으로 돌아가기: b / 메인으로 돌아가기: h ");
-                        System.out.println("즐겨찾기에 추가: f / 댓글 등록: c ");
-                        System.out.print("명령어 입력: ");
-                        String detailCommand = sc.next();
-
-                        switch (detailCommand.toLowerCase()) {
-                            case "b":
-                                inDetailView = false; // 상세보기 루프 종료
-                                break;
-                            case "h":
-                                System.out.println(">> 메인 메뉴로 돌아갑니다.");
-                                return;
-
-                            case "f":
-
-
-                            case "c":
-
-                            default:
-                                System.out.println(">> 잘못된 명령어입니다. b, h, f, c 중에서 입력해주세요.");
-                                break;
-
-                        }
-
-
-                    }
-
-
-                } else {
-                    System.out.println(">> 해당 번호의 여행지가 존재하지 않습니다. 목록에서 번호를 확인해주세요.");
-                }
-                // 상세보기가 끝나면 루프가 계속되어 현재 페이지 목록을 다시 보여줍니다
-
-
-            } catch (NumberFormatException e) {
-
-                switch (command.toLowerCase()) {
-                    case "p":
-                    case "P":
-                        if (currentPage > 0) {
-                            currentPage--;
-                        } else {
-                            System.out.println(">> 첫 페이지 입니다.");
-                        }
-                        break;
-                    case "n":
-                    case "N":
-                        if (currentPage < totalPages - 1) {
-                            currentPage++;
-                        } else {
-                            System.out.println(">> 마지막 페이지 입니다.");
-                        }
-                        break;
-                    case "h":
-                    case "H":
-                        System.out.println(">> 메인 메뉴로 돌아갑니다.");
-                        return; // 루프를 탈출하고 메소드를 종료하여 제어권을 반환
-                    default:
-                        System.out.println(">> 잘못된 명령어입니다. p, n, h 중에서 입력해주세요.");
-                        break;
-                }
-
-            }
-
-
-        }
-
-
-    }
-
-    public void displayNoData() {
-    }
-
-    public static void displayTravelDetail(TravelDTO t) {
-        System.out.println("조회된 여행지는 다음과 같습니다.");
-        System.out.println();
-
-        System.out.print("(no: " + t.getNo() + ") ");
-        System.out.println("[" + t.getDistrict() + "] " + t.getTitle());
-        System.out.println("전화번호: " + t.getPhone());
-        System.out.println("주소: " + t.getAddress());
-        System.out.println("--------------------------------------------------------------------------------------------------------------------------------------");
-        printWithWrap(t.getDescription(), 92);
-        System.out.println("--------------------------------------------------------------------------------------------------------------------------------------");
-    }
-
-
-    public static void printWithWrap(String text, int lineLength) {
-        for (int i = 0; i < text.length(); i += lineLength) {
-            int end = Math.min(i + lineLength, text.length());
-            System.out.println(text.substring(i, end));
-        }
-    }
 
 }
