+++ src/com/multi/view/MyPage.java
@@ -2,7 +2,9 @@ package com.multi.view;
 
 import com.multi.controller.EditProfileController;
 import com.multi.controller.LikeController;
+import com.multi.model.dao.UserDAO;
 import com.multi.model.dto.tmddk.User;
+import com.multi.service.GeneralService;
 import com.multi.service.UserSession;
 
 import java.util.InputMismatchException;
@@ -15,7 +17,7 @@ public class MyPage {
 
     LikeController likeController = new LikeController();
     EditProfileView editProfileView = new EditProfileView(new EditProfileController(), this);
-
+    GeneralService generalService = new GeneralService();
